Miles fixed

+++ src/Main.java
@@ -2,7 +2,7 @@ public class Main {
     public static void main(String[] args) {
         int ticketPrice = 35_870;
         int oneMileCost = 20;
-        double totalMilesBonuses = (double)ticketPrice / oneMileCost;
+        int totalMilesBonuses = (int) ticketPrice / oneMileCost;
         System.out.println(totalMilesBonuses);
     }
 }
